% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarquedBezier.R
\name{.arrowhead_grobs}
\alias{.arrowhead_grobs}
\title{Disentangle Arrow's Head from Its Path}
\usage{
.arrowhead_grobs(arrow, data, panel_params, coord, na.rm)
}
\arguments{
\item{arrow}{\code{grid::arrow()} specification.}

\item{data, panel_params, coord, na.rm}{arguments to \code{Geom$draw_panel()}.}
}
\value{
A grid graphics list (\code{gList}) containing grob(s) that represent the arrow head(s).

The function handles distinct cases for drawing arrowheads at the
start ("first"), end ("last"), or both ends of the path ("both"). This
is based on the \code{arrow$ends} parameter, with 1 representing the start, 2
the end, and 3 indicating both.
}
\description{
The \code{.arrow_grobs} function creates arrowheads separately from the segment.
It's useful to customize the linetype of the segment without affecting the
head.
}
\details{
Call this function from within the \verb{$draw_group()} method of a geom.
}
