% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarquedBezier.R
\name{.cap_mask}
\alias{.cap_mask}
\title{Mask Rows Based on Geometric Cap Aesthetics}
\usage{
.cap_mask(coords)
}
\arguments{
\item{coords}{A data frame containing coordinates with columns \code{x}, \code{y},
\code{start_cap}, \code{end_cap}, and \code{group}.}
}
\value{
A logical vector of the same length as the number of rows in \code{coords}.
Each element is \code{TRUE} if the corresponding point is outside the \code{start_cap}
and \code{end_cap}, and \code{FALSE} if it is within one of these caps.
}
\description{
This function generates a logical vector mask for the rows of a data frame
containing coordinates (\code{coords}), excluding those points that fall within
the geometric areas defined by the \code{start_cap} and/or \code{end_cap} aesthetics.
}
\details{
The function is designed to emulate the internal functionality of
\code{ggraph::geom_edges}. Currently, it supports circular and rectangular caps.
Points that lie within the specified circular radius or within the
rectangular bounds will be excluded, i.e., marked as \code{FALSE} in the mask.

The function first checks whether the \code{start_cap} and \code{end_cap} are
provided and not missing. It identifies the cap type using an internal
function \code{ggraph:::geo_type}. Based on the cap type (\code{circle} or \code{rect}),
it computes the necessary geometric checks.
}
