% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarquedBezier.R
\docType{data}
\name{GeomMarquedPath}
\alias{GeomMarquedPath}
\alias{geom_marqued_bezier}
\title{Create a Path labelled with Marquee}
\format{
An object of class \code{GeomMarquedPath} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomMarquedPath

geom_marqued_bezier(
  mapping = NULL,
  data = NULL,
  stat = "bezier_segment",
  ...,
  position = "identity",
  arrow = NULL,
  na.rm = FALSE,
  show.label = TRUE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\value{
A ggplot2 layer that draws paths with enhanced features such as arrows with always solid lines and intelligently placed labels.
}
\description{
Create paths with optional labels that use marquee styling. The label is
automatically positioned at the midpoint of the path, following its
orientation. This geometry also decouples the arrow's linestyle from the
path's linestyle: in normal geom_path, the arrowhead's linetype follows
the dashing or dotting of the path, which is sometimes confusing.
}
\section{Aesthetics}{

\code{StatBezierSegment} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{x}: The x-coordinate of points defining the path (required).
\item \strong{y}: The y-coordinate of points defining the path (required).
\item \code{alpha}: Transparency level of the path.
\item \code{color}: Color of the path.
\item \code{linetype}: Line type for the path. Note that the linetype for the arrow is always solid for clarity.
\item \code{linewidth}: Line width of the path.
\item \code{label}: Text label for the midpoint of the path.
\item \code{label_alpha}: Transparency level of the label.
\item \code{label_color}: Color of the label.
\item \code{label_fill}: Fill color for the label's background.
\item \code{label_angle}: Angle of the label text, by default automatically aligned with the path direction.
\item \code{label_family}: Font family of the label text.
\item \code{label_fontface}: Font face of the label text (e.g., bold or italic).
\item \code{label_lineheight}: Line height of the label.
\item \code{label_size}: Text size for the label.
\item \code{label_hjust}: Horizontal justification of label text.
\item \code{label_vjust}: Vertical justification of label text.
\item \code{label_nudge}: Numerical value to nudge the label perpendicular to the path direction.
\item \code{label_style}: Style of the marquee label.
\item \code{label_width}: Width of the label's bounding box.
\item \code{start_cap}: Style for the start cap, similar to those in the \code{ggraph} package.
\item \code{end_cap}: Style for the end cap, similar to those in the \code{ggraph} package.
}
}

\section{Parameters}{

Geom parameters include:
\itemize{
\item \code{na.rm}: Logical indicating whether missing values should be removed.
\item \code{arrow}: Specification for arrow heads created using the \code{grid::arrow} function.
\item \code{show.label}: Logical to control whether the label should be drawn.
}
}

\keyword{datasets}
