% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarquedBezier.R
\docType{data}
\name{StatBezierSegment}
\alias{StatBezierSegment}
\alias{stat_bezier_segment}
\title{Create Bezier Curves between Define Points}
\format{
An object of class \code{StatBezierSegment} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 6.

An object of class \code{ggproto} inheriting from \code{Stat}.
}
\usage{
StatBezierSegment

stat_bezier_segment(
  mapping = NULL,
  data = NULL,
  geom = "marqued_bezier",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  n = 100,
  inherit.aes = TRUE,
  ...
)
}
\description{
This stat computes Bezier curves between specified points in a ggplot.
It offers control over the curvature and style of the arcs through
optional aesthetics, enhancing the visualization of paths or connections
in data.
}
\section{Aesthetics}{

\code{StatBezierSegment} understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{x}: Starting x-coordinate of the segment.
\item \strong{y}: Starting y-coordinate of the segment.
\item \strong{xend}: Ending x-coordinate of the segment.
\item \strong{yend}: Ending y-coordinate of the segment.
\item \code{angle}: Angle of the initial tangent vector relative to the segment in degrees, default is 0.
\item \code{strength}: Distance from the start point to the control point, default is half the segment length.
\item \code{angle_end}: Angle of the terminating tangent vector relative to the segment in degrees, default is the same as \code{angle}.
\item \code{strength_end}: Distance from the end point to the control point, default is the same as \code{strength}.
}
}

\section{Dropped Aesthetics}{

The following columns are removed from the data frame passed to the compute function:
\itemize{
\item \code{xend}
\item \code{yend}
\item \code{angle}
\item \code{strength}
\item \code{angle_end}
\item \code{strength_end}
}
}

\section{Extra Parameters}{

The following parameters can be set via \code{params}:
\itemize{
\item \code{na.rm}: A logical indicating whether missing values should be removed.
\item \code{n}: The number of interpolated points to create the Bezier curve.
}
}

\section{Computation}{

The computation involves the following steps:
\enumerate{
\item Compute the length and angle of each segment using the Euclidean distance and \code{atan2}.
\item Set the controls for the Bezier curves, either using provided values or defaults.
\item Calculate control points and define Bezier path using a custom utility to generate paths.
\item Generate a data frame comprising the interpolated path points, where one path is formed per segment.
}
}

\examples{
\dontrun{
library(ggplot2)
data <- data.frame(
  x = c(1, 2),
  y = c(1, 2),
  xend = c(3, 4),
  yend = c(2, 3)
)
ggplot(data) +
  geom_path(aes(x = x, y = y, group = group), stat = "BezierSegment")
}

}
\seealso{
\code{\link[ggforce]{getBeziers}} for details on Bezier calculations.

\code{\link{geom_marqued_bezier}}.
}
\keyword{datasets}
